+++ test/test_helper.rb
@@ -3,6 +3,7 @@ ENV["RAILS_ENV"] = "test"
 require_relative "dummy/config/environment"
 ActiveRecord::Migrator.migrations_paths = [ File.expand_path("dummy/db/migrate", __dir__) ]
 require "rails/test_help"
+require "tmpdir"
 
 ActiveSupport::TestCase.fixture_paths = [ File.expand_path("fixtures", __dir__) ]
 ActionDispatch::IntegrationTest.fixture_paths = ActiveSupport::TestCase.fixture_paths